% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ci.R
\name{calculate_ci}
\alias{calculate_ci}
\title{Compute confidence interval bounds from confidence level}
\usage{
calculate_ci(conf.level = 0.95)
}
\arguments{
\item{conf.level}{Confidence level, expressed as a proportion in (0, 1); default is 0.95.}
}
\description{
Compute confidence interval bounds from confidence level
}
