% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_incidence.R
\name{calculate_incidence}
\alias{calculate_incidence}
\title{Calculate incidence rates}
\usage{
calculate_incidence(
  linked_data,
  type,
  id_col = "id",
  date_col = "date",
  pop_data = NULL,
  pop_col = "pop_count",
  person_time_data = NULL,
  person_time_col = NULL,
  time_p = NULL,
  grouping_vars = NULL,
  only_counts = FALSE,
  suppression = TRUE,
  suppression_treshold = 5,
  log_path = NULL
)
}
\arguments{
\item{linked_data}{Dataset containing relevant diagnostic and demographic information}

\item{type}{Can either be cumulative or rate}

\item{id_col}{Name (character) of the ID column in the data set (unique personal identifier). Default is "id".}

\item{date_col}{Name (character) of the date column in the data set. Default is "date".}

\item{pop_data}{Dataset containing relevant population information.}

\item{pop_col}{Name (character) of the column containing population counts in the population dataset.}

\item{person_time_data}{Dataset containing relevant person-time information.}

\item{person_time_col}{Name (character) of the column containing person-time counts in the person-time dataset.}

\item{time_p}{Time period or time point. For time period, specify as a range. For time point, single numerical value.}

\item{grouping_vars}{Optional character vector including grouping variables for the aggregation of diagnostic counts (eg. sex, education).}

\item{only_counts}{Return only diagnostic count, instead of prevalence rates. Default is set to FALSE.}

\item{suppression}{Apply suppression to results (intermediate and rates) in order to maintain statistical confidentiality.}

\item{suppression_treshold}{Threshold for suppression, default is set to 5 (NPR standard).}

\item{log_path}{File path log}
}
\value{
Prevalence rate table
}
\description{
The \code{calculate_incidence()} function calculates incidence rates based on the given diagnostic and demographic information.
Incidence represents the number of new cases of a given diagnosis that exist in a population of interest at a specified point or period in time.
}
