% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prev.R
\name{calculate_prev}
\alias{calculate_prev}
\title{Calculate prevalence rates}
\usage{
calculate_prev(
  linked_data,
  id_col = "id",
  date_col = "date",
  pop_data,
  pop_col = "pop_count",
  time_p = NULL,
  grouping_vars = NULL,
  only_counts = FALSE,
  suppression = TRUE,
  suppression_treshold = 5
)
}
\arguments{
\item{linked_data}{Dataset containing relevant diagnostic and demographic information}

\item{id_col}{Name (character) of the ID column in the data set (unique personal identifier). Default is "id".}

\item{pop_data}{Dataset containing relevant population information.}

\item{pop_col}{Name (character) of the column containing population counts in the population dataset.}

\item{time_p}{Optional time period or time point. Useful to calculate either point or period prevalence.}

\item{grouping_vars}{Optional character vector including grouping variables for the aggregation of diagnostic counts (eg. sex, education).}

\item{only_counts}{Return only diagnostic count, instead of prevalence rates. Default is set to FALSE.}

\item{suppression}{Apply suppression to results (intermediate and rates) in order to maintain statistical confidentiality.}

\item{suppression_treshold}{Threshold for suppression, default is set to 5 (NPR standard).}
}
\value{
Prevalence rate table
}
\description{
The `calculate_prev()` function calculates prevalence rates based on the given diagnostic and demographic information.
Prevalence represents the number of cases of a given diagnosis that exist in a population of interest at a specified point or period in time.
}
