% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prevalence.R
\name{calculate_prevalence}
\alias{calculate_prevalence}
\title{Calculate prevalence rates}
\usage{
calculate_prevalence(
  linked_data,
  id_col = "id",
  date_col = "date",
  pop_data = NULL,
  pop_col = "pop_count",
  time_p,
  grouping_vars = NULL,
  CI = T,
  only_counts = FALSE,
  suppression = TRUE,
  suppression_treshold = 5,
  log_path = NULL
)
}
\arguments{
\item{linked_data}{Dataset containing relevant diagnostic and demographic information}

\item{id_col}{Name (character) of the ID column in the data set (unique personal identifier). Default is "id".}

\item{date_col}{Name (character) of the date column in the data set. Default is "date".}

\item{pop_data}{Dataset containing relevant population information.}

\item{pop_col}{Name (character) of the column containing population counts in the population dataset.}

\item{time_p}{Time period or time point. For time period, specify as a range. For time point, single numerical value. Useful to calculate either point or period prevalence.}

\item{grouping_vars}{Optional character vector including grouping variables for the aggregation of diagnostic counts (eg. sex, education).}

\item{CI}{Optional computation of confidence intervals}

\item{only_counts}{Return only diagnostic count, instead of prevalence rates. Default is set to FALSE.}

\item{suppression}{Apply suppression to results (intermediate and rates) in order to maintain statistical confidentiality.}

\item{suppression_treshold}{Threshold for suppression, default is set to 5 (NPR standard).}

\item{log_path}{File path of the log file to be used}

\item{CI_level}{Level for confidence intervals, default is set to .99}
}
\value{
Prevalence rate table
}
\description{
The `calculate_prevalence()` function calculates prevalence rates based on the given diagnostic and demographic information.
Prevalence represents the number of cases of a given diagnosis that exist in a population of interest at a specified point or period in time.
}
