% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prevalence_series.R
\name{calculate_prevalence_series}
\alias{calculate_prevalence_series}
\title{Calculate prevalence series rates}
\usage{
calculate_prevalence_series(
  linked_data,
  time_points,
  id_col = "id",
  date_col = "date",
  pop_data,
  pop_col = "pop_count",
  grouping_vars = NULL,
  only_counts = FALSE,
  suppression = TRUE,
  suppression_threshold = 5,
  CI = TRUE,
  CI_level = 0.99,
  log_path = NULL
)
}
\arguments{
\item{linked_data}{A data frame containing linked relevant diagnostic and demographic information.}

\item{time_points}{A list containing either individual time points or time period (range).
\itemize{
\item For time points, each element of the list should be an individual year. For example, \code{time_points = list(2012, 2013, 2014)}
\item For time periods, each element of the list should have two years representing the range of years in the desired period. For example, \code{time_points <- list(c(2012,2014), c(2014,2016), c(2016,2018))}
}}

\item{id_col}{A character string. Name of ID (unique personal identifier) column in \code{linked_data}. Default is "id".}

\item{date_col}{A character string. Name  of the date column in \code{linked_data}. Default is "date".}

\item{pop_data}{A data frame containing corresponding population count information.}

\item{pop_col}{A character string. Name of the column containing population counts in \code{pop_data}.}

\item{grouping_vars}{Character vector (optional). Grouping variables for the aggregation of diagnostic counts (e.g. sex, education).}

\item{only_counts}{Logical. Only want diagnostic counts? Default is \code{FALSE}.
\itemize{
\item If \code{TRUE}, return only counts.
}}

\item{suppression}{Logical. Suppress results (counts and rates) in order to maintain statistical confidentiality? Default is \code{TRUE}.
\itemize{
\item If \code{TRUE}, applies primary suppression (NA) to any value under the threshold defined by \code{suppression_threshold}
}}

\item{suppression_threshold}{Integer. Threshold used for suppression, default is set to 5 (NPR standard).}

\item{CI}{Logical. Want to compute binomial confidence intervals? Default is \code{TRUE}.
\itemize{
\item If \code{TRUE}, add two new columns with the upper and lower CI bound with significance level defined by \code{CI_level}. Uses the Pearson-Klopper method.
}}

\item{CI_level}{A numerical value between 0 and 1. Level for confidence intervals, default is set to 0.99}

\item{log_path}{A character string. Path to the log file to append function logs. Default is \code{NULL}.
\itemize{
\item If \code{NULL}, a new directory \verb{/log} and file is created in the current working directory.
}}
}
\value{
Prevalence series for specified time points/periods
}
\description{
The \code{calculate_prevalence_series()} function calculates prevalence rates series based on the given diagnostic and demographic information. Use \code{calculate_prevalence()} function for only one time period or time point.
Prevalence represents the number of cases of a given diagnosis that exist in a population of interest at a specified point or period in time.
}
\examples{

log_file <- tempfile()
cat("Example log file", file = log_file)

pop_df <- tibble::tibble(year = c(2012:2020), population = floor(runif(9, min=3000, max=4000)))
linked_df <- linked_df |> dplyr::rename("year"= "y_diagnosis_first")

prevalence_df <- calculate_prevalence_series(linked_df,
  time_points = list(2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020),
  id_col = "id",
  date_col = "year",
  pop_data = pop_df,
  pop_col = "population",
  only_counts = FALSE,
  suppression = TRUE,
  suppression_threshold = 1,
  CI = TRUE,
  CI_level = 0.95,
  log_path = log_file)

}
