% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coefplot}
\alias{coefplot}
\title{Graphically compare sizes of regression coefficient}
\usage{
coefplot(reg, data = NULL, sort = TRUE, remove.intercept = TRUE,
  show.units = !is.null(data))
}
\description{
mainly useful if regression is run on data that has been normalized by set.data.units
}

