% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_diag.R
\name{count_diag}
\alias{count_diag}
\title{Count diagnostic cases}
\usage{
count_diag(data, grouping_vars = NULL, id_col = "id")
}
\arguments{
\item{data}{Minimal tidy dataset containing both diagnostic and demographic information.}

\item{grouping_vars}{Character vector including grouping variables for the aggregation of diagnostic counts (eg. sex, education).}

\item{id_col}{Name (character) of the ID column in the data set (unique personal identifier). Default is "id".}
}
\value{
Dataset containing the diagnostic counts with the specified grouping
}
\description{
Count diagnostic cases
}
