\name{courseRecords}
\alias{courseRecords}

\title{
Records from several offerings of a certain course.
}

\description{
The data are in the form of an R list.  Each element of the list
corresponds to one offering of the course.  Fields are:  Class level;
major (two different computer science majors, LCSI in Letters and
Science and ECSE in engineering); quiz grade average (scale of 4.0, A+
counting as 4.3); homework grade average (same scale); and course letter
grade.
}


