% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_diag.R
\name{curate_diag}
\alias{curate_diag}
\title{Curate diagnostic data}
\usage{
curate_diag(
  data,
  min_diag = 1,
  first_diag = TRUE,
  id_col = "id",
  code_col = "icd_code",
  date_col = "date",
  log_path = NULL
)
}
\arguments{
\item{data}{A data frame containing pre-processed and validated diagnostic data.}

\item{min_diag}{Integer. Number of minimum amount of diagnostic events. Default is 1.}

\item{first_diag}{Logical. If \code{TRUE}, keep only information from the first recorded diagnostic event. Default set to TRUE. Default is \code{TRUE}.}

\item{id_col}{A character string. Name of ID column in \code{data}, default is "id"}

\item{code_col}{A character string. Name of column containing the ICD-10 codes in \code{data}, default is "icd_code"}

\item{date_col}{A character string. Name of column containing the diagnostic date in \code{data}, default is "date"}

\item{log_path}{A character string. Path to the log file to append function logs. Default is \code{NULL}.
\itemize{
\item If \code{NULL}, a new directory \verb{/log} and file is created in the current working directory.
}}
}
\value{
Curated diagnostic data: minimum diagnostic events, and/or first ever diagnosis information
}
\description{
Curate diagnostic data
}
\examples{
# Keep only curated diagnostic data
# for example minimum diagnostic events or first recorded diagnosis

log_file <- tempfile()
cat("Example log file", file = log_file)

curated_diag_df <- curate_diag(data = diag_df,
                               min_diag = 1,
                               first_diag = TRUE,
                               id_col = "id",
                               code_col = "code",
                               date_col = "diag_year",
                               log_path = log_file)
}
