\name{day,day1}
\alias{day}
\alias{day1}

\title{
Bike sharing data.
}

\description{
This is the Bike Sharing dataset (day records only) from the UC Irvine
Machine Learning Dataset Repository.  Included here with 
permission of Dr. Hadi Fanaee.  

The \code{day} data is as on UCI; \code{day1} is modified so that the
numeric weather variables are on their original scale.

See \url{https://archive.ics.uci.edu/ml/datasets/bike+sharing+dataset}
for details.
}


