% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_df.R
\docType{data}
\name{diag_df}
\alias{diag_df}
\title{Example diagnostic individual-level data}
\format{
\code{diag_df}
A data frame with 30,024 rows and 3 columns:
\describe{
\item{id}{Unique personal identifier}
\item{code}{ICD-10 codes}
\item{diag_year}{Year of diagnosis}
}
}
\source{
Code used to simulate this dataset can be found in /data-raw directory
}
\usage{
diag_df
}
\description{
A simulated dataset including individual information such as unique ID, ICD-10 and diagnosis date (year).
}
\keyword{datasets}
