% Generated by roxygen2 (4.0.1): do not edit by hand
\name{effectplot}
\alias{effectplot}
\title{Plot for regressions to compare effects sizes of normalized changes in the explanatory variables}
\usage{
effectplot(reg, dat, vars = intersect(colnames(dat), names(coef(reg))),
  ignore.vars = NULL, numeric.effect = "10-90", dummy01 = TRUE,
  sort = TRUE, scale.depvar = NULL, depvar = names(reg$model)[[1]],
  xlab = "Explanatory variables\\n(low baseline high)",
  ylab = paste0("Effect on ", depvar, ""), colors = c(pos = "#11AAAA", neg =
  "#EE3355"), effect.sizes = NULL, effect.bases = NULL, horizontal = TRUE,
  show.ci = FALSE, ci.prob = c(0.05, 0.95), ...)
}
\arguments{
\item{reg}{the results from a regression, e.g. from a call to lm or glm}

\item{dat}{the data frame the regression was estimated from}

\item{vars}{the explanatory variables that shall be shown in the regression lplot}

\item{numeric.effect}{a code describing the lowest and highest values of numeric explanatory variables used to calculate the effect, e.g. "05-95" means taking the effect of moving from the 5% to the 95% quantile.}

\item{dummy01}{shall numeric varibles that have only 0 and 1 as values be treated as a dummy variables?}

\item{sort}{if TRUE (default) sort the effects by size}

\item{scale.depvar}{}

\item{depvar}{name of the dependent variable}

\item{xlab,}{ylab labels}

\item{colors}{colors for positive values (pos) and negative values (neg)}

\item{horizontal}{shall bars be shown horizontally}

\item{show.ci}{shall confidence intervals be shown}

\item{ci.prob}{left and right probability level for confidence intervals}
}
\description{
The plot shall help to compare magnitudes of the influence of different explanatory variables. The default effect is "10-90", i.e. the effect of when -ceteris paribus- changing an (numeric) explanatory variable from its 10% quantile value to its 90% quantile. For dummy variables, we just consider the effect from changing it from 0 to 1.
}

