% Generated by roxygen2 (4.0.1): do not edit by hand
\name{expanded.regression}
\alias{expanded.regression}
\title{Rerun the regression reg with a formula and data in which all factors and transformations are expanded to separate (dummy-)variables.}
\usage{
expanded.regression(reg, org.formula = formula(reg),
  org.data = get.regression.data(reg), parent.env = parent.frame())
}
\arguments{
\item{reg}{the original regression object (needs to have a fields call, formula and data)}
}
\description{
Rerun the regression reg with a formula and data in which all factors and transformations are expanded to separate (dummy-)variables.
}

