% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_diag.R
\name{filter_diag}
\alias{filter_diag}
\title{Validate and filter diagnostic data by selected ICD-10 codes}
\usage{
filter_diag(
  data,
  codes = NULL,
  pattern_codes = NULL,
  id_col = "id",
  code_col = "icd_code",
  log_path = NULL
)
}
\arguments{
\item{data}{A data frame containing pre-processed diagnostic data.}

\item{codes}{A character vector. ICD-10 codes to validate and filter in \code{data}}

\item{pattern_codes}{A character vector. Pattern of ICD-10 codes to validate and filter in \code{data}. For example, F84 will use all codes starting with F84 (F840, F841, F842, F844, etc.)}

\item{id_col}{A character string. Name of ID column in \code{data}, default is "id"}

\item{code_col}{A character string. Name of column containing the ICD-10 codes in \code{data}, default is "icd_code"}

\item{log_path}{A character string. Path to the log file to append function logs. Default is \code{NULL}.
\itemize{
\item If \code{NULL}, a new directory \verb{/log} and file is created in the current working directory.
}}
}
\value{
Filtered and validated diagnostic data frame containing only relevant observations based on diagnostic codes of interest.
}
\description{
Validate and filter diagnostic data by selected ICD-10 codes
}
\examples{
# Validate that F45 and F84 are real codes/family of codes in ICD-10.
# Keep only rows with codes containing F45 and F84.

log_file <- tempfile()
cat("Example log file", file = log_file)

filtered_diag_df <-  filter_diag(data = diag_df,
                                 pattern_codes = c("F45", "F84"),
                                 id_col = "id",
                                 code_col = "code",
                                 log_path = log_file
                                 )

}
