% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.coef.mat}
\alias{get.coef.mat}
\title{get the matrix with coefficients, se, t-value, p-value of a regression model}
\usage{
get.coef.mat(mod, vcov = NULL, robust = FALSE, robust.type = "HC3",
  coef.transform = NULL, cluster1 = NULL, cluster2 = NULL,
  coef.mat = NULL)
}
\description{
internally used by showreg
}

