% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.regression.data}
\alias{get.regression.data}
\title{Extract data from a regression model}
\usage{
get.regression.data(mod)
}
\description{
First looks whether the model mod has a field 'custom.data' in which a user can manually store the data.frame.
}

