% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_ssb.R
\name{get_population_ssb}
\alias{get_population_ssb}
\title{Get population data from SSB}
\usage{
get_population_ssb(
  url_api = "https://data.ssb.no/api/v0/en/table/07459/",
  regions = "norway",
  years,
  ages,
  aggregate_age = TRUE,
  by_sex = TRUE,
  save_xslx = FALSE,
  log_path = NULL
)
}
\arguments{
\item{url_api}{URL used for the SSB API call. Do not modify.}

\item{regions}{Supported regions: "norway" for the country-wise population counts, "fylker" for county population counts, "kommuner" for municipality population counts. Due to region reforms in Norway, the results are shown using SSB's harmonized county codes.}

\item{years}{Numerical or character vector containing year or years}

\item{ages}{Desired ages in whole years}

\item{aggregate_age}{If TRUE and more than one age given, new row with aggregated population for age group}

\item{by_sex}{Should population results be disaggregated by sex}

\item{save_xslx}{Want to save the results as a xslx in your current working directory}

\item{log_path}{Log path}
}
\value{
Data frame with population data from SSB
}
\description{
Get population data from SSB
}
\examples{
# Population of Norway in 2020 to 2022, for ages 10 to 15, include total of that age group (10-15) for each year.
log_file <- tempfile()
cat("Example log file", file = log_file)
population_norway <- get_population_ssb(regions = "norway",
                                        years = c(2020:2022),
                                        ages = c(10:15),
                                        aggregate_age = TRUE,
                                        by_sex = TRUE,
                                        save_xslx = FALSE,
                                        log_path = log_file)

}
