% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_municipality_codes.R
\name{harmonize_municipality_codes}
\alias{harmonize_municipality_codes}
\title{Harmonize old municipality codes in Norway to 2024 municipality codes}
\usage{
harmonize_municipality_codes(data, municipality_col = "code", fylke = FALSE)
}
\arguments{
\item{data}{Data frame, or data frame extensions (e.g. tibble).}

\item{municipality_col}{Character vector. Name of column containing the original municipality codes.}

\item{fylke}{Logical. If \code{TRUE}, output data frame also includs a column with the name of the corresponding fylke in 2024. Default is \code{FALSE}.}
}
\value{
Data frame with the old and equivalent municipal codes in 2024.
}
\description{
Harmonize old municipality codes in Norway to 2024 municipality codes
}
\examples{
# Harmonize municipality codes from 2016 to 2024

harmonized_codes <- harmonize_municipality_codes(data = kommuner_2016, municipality_col = "code")
}
