% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invar_df.R
\docType{data}
\name{invar_df}
\alias{invar_df}
\title{Example time-invariant individual-level data}
\format{
`invar_df`
A data frame with 30,024 rows and 4 columns:
\describe{
  \item{id}{Unique personal identifier}
  \item{sex}{Sex (code)}
  \item{y_birth}{Year of birth}
  \item{innvandringsgrunn}{Reason for immigration (https://www.ssb.no/klass/klassifikasjoner/355)}
}
}
\source{
Code used to simulate this dataset can be found in /data-raw directory
}
\usage{
invar_df
}
\description{
A simulated dataset including time-invariant individual information such as year of residence and place of residence (codes)
}
\keyword{datasets}
