% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kommuner_2016.R
\docType{data}
\name{kommuner_2016}
\alias{kommuner_2016}
\title{Norwegian municipalities valid in 2016}
\format{
\code{kommuner_2016}
A data frame with 429 rows and 2 columns:
\describe{
\item{code}{Code of municipality (2016)}
\item{name}{Name of municipality}
}
}
\source{
Code used to simulate this dataset can be found in /data-raw directory
}
\usage{
kommuner_2016
}
\description{
Data frame including Norwegian municipality codes valid in 2016.
}
\keyword{datasets}
