% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincom.R
\name{lincom}
\alias{lincom}
\title{Linear combination of regression coefficients}
\usage{
lincom(model, weights, vcov. = NULL)
}
\arguments{
\item{model}{A regression model with \code{coef} and \code{vcov} methods.}

\item{weights}{A numeric vector of coefficient weights to construct the
combination, or a matrix in which each row represents one such
combination. Each element specifies the weight applied to the
corresponding element of \code{attr(terms(model), "term.labels")}.}

\item{vcov.}{(Optional) Either a function for estimating the covariance
matrix of the regression coefficients, or the covariance matrix itself.
If not specified, the default \code{vcov} method is used.}
}
\description{
This function calculates the point estimate and standard error for a linear
combination of regression coefficients (similar to Stata's "lincom"
function). At this stage the combination must be specified as a numeric
vector (or matrix) of weights -- future updates will extend this to allow
more flexible methods of specifying coefficient combinations (similar to
\code{car}'s \code{linearHypothesis} function).
}
