% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked_df.R
\docType{data}
\name{linked_df}
\alias{linked_df}
\title{Example linked individual-level data}
\format{
\code{linked_df}
A data frame with 30,024 rows and 3 columns:
\describe{
\item{id}{Unique personal identifier}
\item{code}{ICD-10 codes}
\item{diag_year}{Year of diagnosis}
\item{sex}{Sex (code)}
\item{y_birth}{Year of birth}
\item{innvandringsgrunn}{Reason for immigration (https://www.ssb.no/klass/klassifikasjoner/355)}
}
}
\source{
Code used to simulate this dataset can be found in /data-raw directory
}
\usage{
linked_df
}
\description{
A simulated dataset including individual information such as unique ID, ICD-10 and diagnosis date (year), year of birth, and reason for immigration
}
\keyword{datasets}
