% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtools.R
\name{model.dep.var}
\alias{model.dep.var}
\title{Returns the name of the dependent variable of a model object}
\usage{
model.dep.var(mod)
}
\description{
Returns the name of the dependent variable of a model object
}

