% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtools.R
\name{name.of.depvar}
\alias{name.of.depvar}
\title{Try to extract name of the dependent variable from a regression model}
\usage{
name.of.depvar(reg)
}
\description{
Try to extract name of the dependent variable from a regression model
}

