% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rates.R
\name{plot_rates}
\alias{plot_rates}
\title{Title}
\usage{
plot_rates(
  data,
  date_col,
  rate_col,
  grouping_var = NULL,
  facet_var = NULL,
  plot_type = c("line", "bar_chart", "lollipop", "jitter", "slope"),
  percent = TRUE,
  palette = "fhi_colors",
  annotate_line = NULL,
  labels,
  start_end_points = NULL,
  interactive = FALSE
)
}
\arguments{
\item{interactive}{}
}
\description{
Title
}
