% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felm.r
\name{predict.felm}
\alias{predict.felm}
\title{An implementaion of predict for felm}
\usage{
\method{predict}{felm}(object, newdata, use.fe = TRUE, ...)
}
\description{
An implementaion of predict for felm
}

\examples{\dontrun{
  T = 12
  x <- rnorm(T)
  fe1 = sample(c("yes","no"), T, replace=TRUE)
  fe2 = sample(c("a","b","c"), T, replace=TRUE)
  y <- x1  + (fe1=="yes") + 2*(fe2=="a") + (fe2=="b")+ rnorm(T)
  
  new <- data.frame(
    x = seq(-3, 3, length=T), 
    fe1 = sample(c("yes","no"), T, replace=TRUE),
    fe2 = sample(c("a","b","c"), T, replace=TRUE)    
  )
  p.lm = predict(lm(y ~ x+fe1+fe2), new)

  library(lfe)  
  p.felm = predict(felm(y~x|fe1+fe2),new)
  cbind(p.lm, p.felm)
}}
