\name{quizDocs}
\alias{quizDocs}

\title{
Course quiz documents
}

\description{

This data is suitable for NLP analysis.

This is an R list, 143 elements, one for each of 143 quizzes from my
various courses.  Each list element is a character vector, one vector
element per line of the quiz.

The original documents were LaTeX files.  They have been run through the
\code{detex} utility to remove most LaTeX commands, as well as removing
the LaTeX preambles separately.

The names of the list elements are the course names, as follows:

ECS 50:  a course in machine organization

ECS 132:  an undergraduate course in probabilistic modeling

ECS 145:  a course in scripting languages (Python, R)

ECS 158:  an undergraduate course in parallel computation

ECS 256:  a graduate course in probabilistic modeling

}

