% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_demographic.R
\name{read_demo_data}
\alias{read_demo_data}
\title{Read and validate the structure of demographic data}
\usage{
read_demo_data(file_path, data_type, id_col = "id", date_col = "date", ...)
}
\arguments{
\item{file_path}{File path of the demographic data to be read. Supports CSV, RDS, RDA and .SAV files.}

\item{data_type}{Demographic data can either be of type "t_variant" or "t_invariant", necessary to check correct data structure characteristics.}

\item{id_col}{Name of ID column in data set, default is "id"}

\item{date_col}{Name of date column in data set, default is "date"}

\item{...}{Optional extra parameters for specifying correct reading of CSV and .SAV files}
}
\value{
A data frame with the validated minimum requirements for demographic data
}
\description{
`read_data_demographic()` validates the general structure and minimum column requirements for demographic data.
The input data sets must be CSV, RDS, RDA or .SAV files.
}
