% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_anatomy.R
\name{regression.anatomy.plots}
\alias{regression.anatomy.plots}
\title{Regression anatomy plots}
\usage{
regression.anatomy.plots(mod, vars = NULL)
}
\arguments{
\item{mod}{A regression model, a result of a call to lm.}

\item{vars}{which x variables shall be included in the plot}
}
\description{
Show original scatterplots and scatterplots
of y and controlled x_i, i.e. the residuals
of a regression of x_i on the other explanatory
variables. See e.g. Filoso (2010):
https://mpra.ub.uni-muenchen.de/42716/1/FILOSO-Regression_Anatomy_Revealed.pdf
}

