% Generated by roxygen2 (4.0.1): do not edit by hand
\name{regression.data}
\alias{regression.data}
\title{Creates a data.frame with all dummy variables and transformed variables described in formula}
\usage{
regression.data(formula, data = parent.frame(), normalize.names = TRUE,
  remove.intercept = TRUE)
}
\arguments{
\item{formula}{a regression formula e.g. y~log(x1)+x2}

\item{data}{a data frame that contains original variables used in formula}

\item{normalize.names}{if TRUE (default) change column names to be valid R variable names}

\item{remove.intercept}{if TRUE (default) do not generate an .Intercept column}
}
\value{
a data.frame
}
\description{
Creates a data.frame with all dummy variables and transformed variables described in formula
}

