% Generated by roxygen2 (4.0.1): do not edit by hand
\name{regression.formula}
\alias{regression.formula}
\title{generates a regression formula from a data.frame assuming the first column is regressed on all other columns}
\usage{
regression.formula(data, lhs = colnames(data)[1],
  rhs = setdiff(colnames(data), lhs))
}
\description{
generates a regression formula from a data.frame assuming the first column is regressed on all other columns
}

