
\name{regtools-package}
\alias{regtools-package}
\alias{regtools}
\docType{package}         

\title{Overview and Package Reference Guide}
                                                                                \description{ 

This package provides a broad collection of functions useful for
regression and classification analysis.

}

\section{Function List}{

(under construction)

\bold{Parametric modeling:}

   \itemize{

   \item nonlinear regression:  nlshc

   \item ridge regression:  ridgelm, plot

   }

\bold{Diagnostic plots:}

   \itemize{

   \item regression diagnostics:  parvsnonparplot, nonparvsxplot,
   nonparvarplot

   \item other: boundaryplot

   }

\bold{Classification:}

   \itemize{

   \item classadjust, avalogtrn, avalogpred, ovalogtrn,
   ovalogpred, knntrn, predict, pwplot, nonparvsxplot

   }

\bold{Machine learning:}

   \itemize{

   \item k-NN: kNN, kmin, knnest, knntrn, preprocessx, meany, vary, loclin,
   predict, kmin, pwplot, bestKperPoint

   \item neural networks: krsFit,multCol

   \item advanced grid search: fineTuning

   }

\bold{Dummies and R factors:}

   \itemize{

   \item conversion between factors and dummies: dummiesToFactor,
   factorsToDummies, factorToDummies

   }

\bold{Statistics:}

   \itemize{

   \item mm

   }

\bold{Matrix:}

   \itemize{

   \item multCols

   }


\bold{Misc.:}

   \itemize{

   \item scaling:  mmscale, unscale

   \item data frames: catDFRow, tabletofakedf

   }


}

