% Generated by roxygen2 (4.0.1): do not edit by hand
\name{robust.se.summary}
\alias{robust.se.summary}
\title{Show a model summary with robust standard errors}
\usage{
robust.se.summary(m, type = "HC3", ...)
}
\description{
Using the sandwich package, creates summary with heteroscedasticity consistent standard errors (Stata´s robust dafault (Stata (2014)))
}

