% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.r
\name{scale.data.cols}
\alias{scale.data.cols}
\title{Change units of columns in a data frame, e.g. measure relative to one standard deviation}
\usage{
\method{scale}{data.cols}(dat, unit = "sd", cols = setdiff(colnames(dat),
  exclude), exclude = NULL, dummy01 = TRUE)
}
\description{
useful to make coefficients in regressions better comparable
}

