% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showreg.r
\name{showreg}
\alias{showreg}
\title{Show and compare regression results}
\usage{
showreg(l, custom.model.names = names(l), omit.stat = c("F", "ser"),
  robust = FALSE, robust.type = "HC3", cluster1 = NULL, cluster2 = NULL,
  vcov.li = NULL, coef.transform = NULL, coef.mat.li = NULL, digits = 2,
  output = c("text", "html", "latex")[1], output.fun = NULL,
  doctype = FALSE, title = NULL, intercept.bottom = FALSE,
  package = getOption("showreg.package")[1], ...)
}
\arguments{
\item{l}{list of models as in screenreg}

\item{custom.model.names}{custom titles for each model. By default the names of the model list.}

\item{robust}{shall robust standard errors be used? Logical single number or a vector specifying for each model.}

\item{robust.type}{the type of robust standard errors. Can be "HAC", "cluster", "HC1" to "HC4" or "NeweyWest". Can be a vector specifying a type for each model.}

\item{cluster1}{and cluster2 if clustered robust standard errors are used, the name of the variables that shall be clustered by}

\item{coef.transform}{either NULL or a vector containing "no" or "mfx", if an entry is "mfx" we show the marginal effects of the corresponding model.}

\item{coef.mat.li}{for highest flexibility, you can also provide a list that contains for each model a matrix or data.frame as returned by coeftest with the columns: coefficent, se, t-value, p-value.}

\item{output}{either "text", "html" or "latex"}

\item{output.fun}{allows a manual output function, e.g. if one has overloaded the design of screenreg, texreg or htmlreg.}

\item{title}{a string shown as title above the table}

\item{package}{the underlying package for creating the tables, either "texreg" or "stargazer". The current default is texreg but that may change.}

\item{...}{additional parameters for stargazer or screenreg, texreg or htmlreg}

\item{vcov.list}{optional a list of covariance matrices of the coefficients for every model}
}
\description{
The function extends and wraps either stargazer or the screenreg, texreg and htmlreg functions in the texreg package. It allows for robust standard errors (also clustered robust standard errors) and can show marginal effects in glm models.
}

