% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Create simulated diagnostic, time-varying and time-unvarying data}
\usage{
simulate_data(
  population_size,
  prefix_ids,
  length_ids,
  family_codes,
  pattern,
  prevalence = NULL,
  diag_years,
  incidence = NULL,
  sex_vector,
  y_birth,
  filler_codes,
  filler_y_birth,
  unvarying_query,
  unvarying_codes = NULL,
  region_level = "kommune",
  region_codes,
  filler_region_codes,
  residence_years,
  date_classifications = NULL
)
}
\arguments{
\item{population_size}{pop}

\item{date_classifications}{}
}
\description{
Create simulated diagnostic, time-varying and time-unvarying data
}
