% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{synthetic_data}
\alias{synthetic_data}
\title{Create synthetic diagnostic, time-varying and time-invariant individual level data}
\usage{
synthetic_data(
  population_size,
  prefix_ids,
  length_ids,
  seed = "123",
  family_codes,
  diag_code_type = "icd",
  pattern = c("increase", "decrease", "random"),
  prevalence = NULL,
  diag_years,
  incidence = NULL,
  sex_vector,
  y_birth,
  filler_codes,
  filler_y_birth,
  invariant_queries = NULL,
  invariant_codes = NULL,
  invariant_codes_filler,
  varying_query = NULL,
  varying_codes = NULL,
  varying_codes_filler = NULL,
  date_classifications = NULL
)
}
\arguments{
\item{population_size}{An integer. Number of total population size (individual).}

\item{prefix_ids}{A character string. Prefix used to construct unique IDs. Default is "P000".}

\item{length_ids}{An integer. Total character length of each ID. Default is 6.}

\item{seed}{A numerical value. Seed used to ensure reproducible results. Default is \code{seed = 123}}

\item{family_codes}{A character vector. Relevant diagnostic (either ICD-10 or ICPC-2 codes or family of codes. Example: \code{family_codes = c("F84", "G")}}

\item{diag_code_type}{A character string. Desired code classification, options are "icd" or "icpc". Default is "icd"}

\item{pattern}{A character string. Pattern of incidence or prevalence rates in simulated data. Possible options are "increase", "decrease" or "random".}

\item{prevalence}{A numeric value between 0 and 1. Prevalence rate expressed as a proportion.}

\item{diag_years}{A numeric vector. Years to be used as relevant diagnostic years.}

\item{incidence}{A numeric value between 0 and 1. Incidence rate expressed as a proportion.}

\item{sex_vector}{A factor or character vector. Factors used to represent sex in the simulated data sets.}

\item{y_birth}{A numeric vector. Years to be used as relevant years of birth.}

\item{filler_codes}{A character vector. Diagnostic codes or family of codes used as fillers. Example: \code{filler_codes = c("R", "P20")}}

\item{filler_y_birth}{A numeric vector. Years to be used as filler years of birth.}

\item{invariant_queries}{A character vector. Uses Statistics Norway API to retrieve desired invariant variable classification(s). Example: \code{invariant_queries = c("innvandringsgrunn")}}

\item{invariant_codes}{Data frame or named list. Codes to be used as relevant invariant codes in dataset.
\itemize{
\item If a data frame is provided, column names will be considered as the names of the invariant variables.
\item If a named list is provided, the name of each element will be consider as the invariant variable name. Example: \code{invariant_codes = list("innvandringsgrunn" = c("ARB", "NRD", "UKJ"), "blodtype" = c("A", "B", "AB", "O"))}
}}

\item{invariant_codes_filler}{Data frame or named list. Codes to be used as filler invariant codes in dataset.}

\item{varying_query}{A character string. Uses Statistics Norway API to retrieve desired varying variable classification(s). Example: \code{varying_query = c("sivilstand")}}

\item{varying_codes}{A character vector. Codes to be used as relevant varying codes in dataset. Example: \code{varying_codes = as.character(0:4)}}

\item{varying_codes_filler}{A character vector. Codes to be used as filler varying codes in dataset. Example: \code{varying_codes_filler = as.character(5:9)}}

\item{date_classifications}{Date used to retrieve classification system from SSB. Format must be \strong{"yyyy-mm-dd"}}
}
\value{
Named list containing two lists. The first list named 'datasets' includes the data frames with individual level diagnostic and sociodemographic data. The second list named 'metadata' includes the exact function call and arguments given by the user
}
\description{
The \code{synthetic_data()} function creates individual-level data sets.
It simulates the structure of diagnostic, time-varying and time-invariant data you might commonly encounter when working with Norwegian medical and sociodemographic data (e.g. NPR and SSB)
}
\examples{
simulated_list <- synthetic_data(
  population_size = 1000,
  prefix_ids = "P000",
  length_ids = 6,
  family_codes = c("F45", "F84"),
  pattern = "increase",
  prevalence = .023,
  diag_years  = c(2012:2020),
  sex_vector = c(0, 1),
  y_birth = c(2010:2018),
  filler_codes = "F",
  filler_y_birth = c(2000:2009),
 invariant_codes = list("innvandringsgrunn" = c("ARB", "NRD", "UKJ")),
  invariant_codes_filler = list("innvandringsgrunn" = c("FAMM", "UTD")),
  varying_query = "fylke"
)
}
