% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_df.R
\docType{data}
\name{var_df}
\alias{var_df}
\title{Example time-variant individual-level data}
\format{
`var_df`
A data frame with 270,216 rows and 3 columns:
\describe{
  \item{id}{Unique personal identifier}
  \item{code}{Code place of residence (time varying)}
  \item{year}{Year of residence}
}
}
\source{
Code used to simulate this dataset can be found in /data-raw directory
}
\usage{
var_df
}
\description{
A simulated dataset including time-variant individual information such as year of residence and place of residence (codes)
}
\keyword{datasets}
