% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vcovRobust}
\alias{vcovRobust}
\title{Wrapper to get robust vcov of different kinds}
\usage{
vcovRobust(model, type = "HC3", cluster1 = NULL, cluster2 = NULL, ...)
}
\arguments{
\item{type}{the type of robust standard errors. Can be "HAC", "cluster", "HC1" to "HC4" or "NeweyWest"}
}
\description{
Wrapper to get robust vcov of different kinds
}

