% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_incidence.R
\name{calculate_incidence}
\alias{calculate_incidence}
\title{Calculate incidence rates}
\usage{
calculate_incidence(
  linked_data,
  type = c("cumulative", "rate"),
  id_col = "id",
  date_col = "date",
  pop_data = NULL,
  pop_col = "pop_count",
  person_time_data = NULL,
  person_time_col = NULL,
  time_p = NULL,
  grouping_vars = NULL,
  only_counts = FALSE,
  suppression = TRUE,
  suppression_threshold = 5,
  CI = TRUE,
  CI_level = 0.99,
  log_path = NULL
)
}
\arguments{
\item{linked_data}{A data frame containing linked relevant diagnostic and demographic information. Should include only first time diagnosis, see 'curate_diag'}

\item{type}{Character string. Valid options are "cumulative" or "rate".
\itemize{
\item If "cumulative",
\item If "rate",
}}

\item{id_col}{A character string. Name of ID (unique personal identifier) column in \code{linked_data}. Default is "id".}

\item{date_col}{A character string. Name  of the date column in \code{linked_data}. Default is "date".}

\item{pop_data}{A data frame containing corresponding population at risk information.}

\item{pop_col}{A character string. Name of the column containing population counts in \code{pop_data}.}

\item{person_time_data}{A data frame containing corresponding person-time information.}

\item{person_time_col}{A character string. Name of the column containing person-time counts in \code{person_time_data}.}

\item{time_p}{A numeric value or numeric vector. Time point or time period used to calculate the incidence.
\itemize{
\item For time period, specify as a range. The first value of the vector is the period's lower bound, and the second element is the period's upper bound. Example:  \code{time_p = c(2010,2015)}
\item For time point, single numeric value. Example: \code{time_p = 2010}
}}

\item{grouping_vars}{Character vector (optional). Grouping variables for the aggregation of diagnostic counts (e.g. sex, education).}

\item{only_counts}{Logical. Only want diagnostic counts? Default is \code{FALSE}.
\itemize{
\item If \code{TRUE}, return only counts.
}}

\item{suppression}{Logical. Suppress results (counts and rates) in order to maintain statistical confidentiality? Default is \code{TRUE}.
\itemize{
\item If \code{TRUE}, applies primary suppression (NA) to any value under the threshold defined by \code{suppression_threshold}
}}

\item{suppression_threshold}{Integer. Threshold used for suppression, default is set to 5 (NPR standard).}

\item{CI}{Logical. Want to compute binomial confidence intervals? Default is \code{TRUE}.
\itemize{
\item If \code{TRUE}, add two new columns with the upper and lower CI bound with significance level defined by \code{CI_level}. Uses the Pearson-Klopper method.
}}

\item{CI_level}{A numerical value between 0 and 1. Level for confidence intervals, default is set to 0.99}

\item{log_path}{A character string. Path to the log file to append function logs. Default is \code{NULL}.
\itemize{
\item If \code{NULL}, a new directory \verb{/log} and file is created in the current working directory.
}}
}
\value{
Incidence table
}
\description{
The \code{calculate_incidence()} function calculates incidence rates based on the given diagnostic and demographic information.
Incidence represents the number of new cases of a given diagnosis that exist in a population of interest at a specified point or period in time.
}
\examples{
log_file <- tempfile()
cat("Example log file", file = log_file)

pop_df <- tibble::tibble(year = "2012-2013", population = 4500)
linked_df <- linked_df |> dplyr::rename("year"= "y_diagnosis_first")

incidence_df <- calculate_incidence(linked_df,
  type = "cumulative",
  id_col = "id",
  date_col = "year",
  pop_data = pop_df,
  pop_col = "population",
  time_p = c(2012,2013),
  only_counts = FALSE,
  suppression = TRUE,
  suppression_threshold = 10,
  log_path = log_file)


}
