% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prev.R
\name{calculate_prev}
\alias{calculate_prev}
\title{Calculate prevalence rates}
\usage{
calculate_prev(
  linked_data,
  id_col = "id",
  pop_data,
  pop_col = "pop_count",
  time_period = NULL,
  grouping_vars = NULL,
  only_counts = FALSE
)
}
\arguments{
\item{linked_data}{Dataset containing relevant diagnostic and demographic information}

\item{id_col}{Name (character) of the ID column in the data set (unique personal identifier). Default is "id".}

\item{pop_data}{Dataset containing relevant population information.}

\item{pop_col}{Name (character) of the column containing population counts in the population dataset.}

\item{time_period}{?? Optional}

\item{grouping_vars}{Optional character vector including grouping variables for the aggregation of diagnostic counts (eg. sex, education).}

\item{only_counts}{Return only diagnostic count, instead of prevalence rates. Default is set to FALSE.}
}
\value{
Prevalence rate table
}
\description{
Calculate prevalence rates
}
