% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_diag.R
\name{curate_diag}
\alias{curate_diag}
\title{Curate diagnostic data}
\usage{
curate_diag(
  data,
  min_diag = 1,
  first_diag = TRUE,
  id_col = "id",
  code_col = "icd_code",
  date_col = "date",
  log_path = NULL
)
}
\arguments{
\item{data}{Data frame containing pre-processed and validated diagnostic data (check minimum requirements in documentation)}

\item{min_diag}{Numerical value, minimum amount of diagnostic events}

\item{first_diag}{Option to summarize information keeping only the first recorded diagnostic event. Default set to TRUE.}

\item{id_col}{Character string containing the name of the ID column in data set, default is "id"}

\item{code_col}{Character string containing the name of the code column in data set, default is "icd_code"}

\item{date_col}{Character string containing the name of date column in data set, default is "date"}

\item{log_path}{File path of the log file to be used}
}
\value{
Curated diagnostic data: minimum diagnostic events, and first ever diagnosis information
}
\description{
Curate diagnostic data
}
\examples{
# Keep only curated diagnositc data, for example minimum diagnostic events or first recorded diagnosis

log_file <- tempfile()
cat("Example log file", file = log_file)

curated_diag_df <- curate_diag(data = filtered_diag,
                               min_diag = 1,
                               first_diag = TRUE,
                               id_col = "id",
                               code_col = "code",
                               date_col = "diag_year",
                               log_path = log_file)
}
