% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_diag.R
\name{curate_diag}
\alias{curate_diag}
\title{Curate diagnostic data}
\usage{
curate_diag(
  data,
  min_diag = 1,
  id_col = "id",
  code_col = "icd_code",
  date_col = "date"
)
}
\arguments{
\item{data}{Data frame containing pre-processed and validated diagnostic data (check minimum requirements in documentation)}

\item{min_diag}{Numerical value, minimum amount of diagnostic events}

\item{id_col}{Character string containing the name of the ID column in data set, default is "id"}

\item{code_col}{Character string containing the name of the code column in data set, default is "icd_code"}

\item{date_col}{Character string containing the name of date column in data set, default is "date"}
}
\value{
Curated diagnostic data: minimum diagnostic events, and first ever diagnosis information
}
\description{
Curate diagnostic data
}
