% Generated by roxygen2 (4.0.1): do not edit by hand
\name{effectplot}
\alias{effectplot}
\title{Plot for regressions to compare effects sizes of normalized changes in the explanatory variables}
\usage{
effectplot(reg, dat = get.regression.data(reg),
  vars = intersect(colnames(dat), names(coef(reg))), ignore.vars = NULL,
  numeric.effect = "10-90", dummy01 = TRUE, sort = TRUE,
  scale.depvar = NULL, depvar = names(reg$model)[[1]],
  xlab = "Explanatory variables\\n(low baseline high)",
  ylab = paste0("Effect on ", depvar, ""), colors = c(pos = "#11AAAA", neg =
  "#EE3355"), effect.sizes = NULL, effect.bases = NULL, horizontal = TRUE,
  show.ci = FALSE, ci.prob = c(0.05, 0.95), num.ticks = NULL,
  add.numbers = TRUE, numbers.align = c("center", "left", "right",
  "left_of_bar_end", "right_of_bar_end")[1], numbers.vjust = ifelse(show.ci,
  0, 0.5), left.margin = 0, right.margin = 0, signif.digits = 2,
  round.digits = 8, alpha = 0.8, ...)
}
\arguments{
\item{reg}{the results from a regression, e.g. from a call to lm or glm}

\item{dat}{default = the data frame the regression was estimated from}

\item{vars}{the explanatory variables that shall be shown in the plot}

\item{numeric.effect}{a code describing the lowest and highest values of numeric explanatory variables used to calculate the effect, e.g. "05-95" means taking the effect of moving from the 5 percent to the 95 percent quantile.}

\item{dummy01}{shall numeric varibles that have only 0 and 1 as values be treated as a dummy variables?}

\item{sort}{if TRUE (default) sort the effects by size}

\item{scale.depvar}{a scaling for the dependent variable}

\item{depvar}{name of the dependent variable}

\item{xlab,}{ylab labels}

\item{colors}{colors for positive values (pos) and negative values (neg)}

\item{horizontal}{shall bars be shown horizontally?}

\item{show.ci}{shall confidence intervals be shown?}

\item{ci.prob}{left and right probability level for confidence intervals}

\item{num.ticks}{the number of tick marks on the effect size axis}

\item{add.numbers}{shall the effect sizes be added as numbers in the plot?}

\item{numbers.align}{How shall the effect size numbers be aligned: "left","center", "right" align all numbers at the same horizontal posistion for all variables. "left_of_bar_end" and "right_of_bar_end" align them at the end of each bar.}

\item{numbers.vjust}{Vertical adjustment of the numbers}

\item{left.margin}{extra margin left of bars as fraction of maximum bar width}

\item{right.margin}{extra margin right of bars as fraction of maximum bar width}

\item{signif.digits}{number of significant digits for effect sizes}

\item{round.digits}{number of digits effect sizes shall be rounded to}

\item{...}{further arguments passed to qplot. E.g. you can set "main" to specify a title of the plot.}
}
\description{
The plot shall help to compare magnitudes of the influence of different explanatory variables. The default effect is "10-90", i.e. the effect of when -ceteris paribus- changing an (numeric) explanatory variable from its 10% quantile value to its 90% quantile. For dummy variables, we just consider the effect from changing it from 0 to 1.
}

\examples{\dontrun{
  # simulate some data
  set.seed(12345)
  n = 1000
  x = rnorm(n)
  z = rnorm(n)
  q = rnorm(n)
  
  # binary outcome
  y = ifelse(pnorm(1 + 0.5*x + 0.25*x^2 - 0.5*z + rnorm(n))>0.5, 1, 0)

  data = data.frame(y,x,z,q)
  # Logit regression
  reg = glm(y~x + x^2 + z +q, data=data, family="binomial")
  effectplot(reg,main="Effects", horizontal=TRUE, show.ci=TRUE)


  # An example with factors
  
  T = 10
  x = sample(1:4, T, replace = TRUE)
  y = x^2 + rnorm(T)
  xf = as.character(x)

  # effectplot can currently not handle factor variables
  # in the regression formula
  reg = lm(y~xf)
  effectplot(reg)
  
  # Workaround: first explicitly generate a 
  # data.frame with all dummy variables
  dat = expanded.regression.data(y~xf)
  reg = lm(regression.formula(dat), data=dat)
  reg
  effectplot(reg)

  # Short-cut
  reg = expanded.regression(lm(y~xf))
  effectplot(reg)
  
}}
