% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_demo.R
\name{filter_demo}
\alias{filter_demo}
\title{Filter demographic data by selected filtering parameters}
\usage{
filter_demo(
  data,
  data_type = c("t_variant", "t_invariant"),
  filter_param,
  id_col = NULL,
  any = FALSE,
  rm_na = TRUE,
  log_path = NULL
)
}
\arguments{
\item{data}{A data frame containing pre-processed demographic data.}

\item{data_type}{A character string. Type of demographic data: "t_variant" or "t_invariant"}

\item{filter_param}{A named list containing filtering parameters. The names in the list are the column names and the values are vectors of values to keep.}

\item{id_col}{A character string. Name of ID column in data set.
\itemize{
\item Optional, necessary only when \code{any = TRUE}
}}

\item{any}{Logical. Filtering option, any year. Default is \code{FALSE}}

\item{rm_na}{Logical. Should rows with NA in the non-filtered columns be removed? Default is \code{FALSE}
\itemize{
\item If \code{TRUE}, removes observations that have NA in any of the non-filtered columns.
}}

\item{log_path}{A character string. Path to the log file to append function logs. Default is \code{NULL}
\itemize{
\item If \code{NULL}, a new directory \verb{/log} and file is created in the current working directory.
}}
}
\value{
Filtered demographic dataframe containing only relevant observations based on the filtering parameters.
}
\description{
Filter demographic data by selected filtering parameters
}
\examples{
# Filter varying and unvarying datasets

log_file <- tempfile()
cat("Example log file", file = log_file)

filtered_var <- filter_demo(data = var_df,
data_type = "t_variant",
filter_param = list("year_varying" = c(2012:2015), "varying_code" = c("03")),
log_path = log_file)


filtered_invar <- filter_demo(data = invar_df, data_type = "t_invariant",
filter_param = list("y_birth" = c(2006:2008),
"innvandringsgrunn" = c("FAM", "UTD")),
rm_na = FALSE,
log_path = log_file)

}
