% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_demo.R
\name{filter_demo}
\alias{filter_demo}
\title{Filter demographic data by selected filtering parameters}
\usage{
filter_demo(
  data,
  data_type = c("t_variant", "t_invariant"),
  filter_param,
  id_col = NULL,
  any = FALSE,
  rm_na = TRUE,
  log_path = NULL
)
}
\arguments{
\item{data}{Data frame containing pre-processed demographic data}

\item{data_type}{Type of demographic data: "t_variant" or "t-variant"}

\item{filter_param}{Named list containing filtering parameters. The names in the list are the column names and the values are vectors of values to keep.}

\item{id_col}{Optional flag, necessary for "any" filtering option.}

\item{any}{Filtering option, any year. Default = FALSE.}

\item{rm_na}{Removes observations that have NA in the non-filtered columns.}

\item{log_path}{File path of the log file to be used}
}
\value{
Filtered demographic dataframe containing only relevant observations based on the filtering parameters.
}
\description{
Filter demographic data by selected filtering parameters
}
\examples{
# Filter varying and unvarying datasets

log_file <- tempfile()
cat("Example log file", file = log_file)

filtered_var <- filter_demo(data = var_df,
data_type = "t_variant",
filter_param = list("year_varying" = c(2012:2015), "varying_code" = c("03")),
log_path = log_file)


filtered_invar <- filter_demo(data = invar_df, data_type = "t_invariant",
filter_param = list("y_birth" = c(2006:2008),
"innvandringsgrunn" = c("FAM", "UTD")),
rm_na = FALSE,
log_path = log_file)

}
