% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_diag_codes.R
\name{filter_diag}
\alias{filter_diag}
\title{Validate and filter diagnostic data by selected ICD-10 codes}
\usage{
filter_diag(
  data,
  codes = NULL,
  pattern_codes = NULL,
  id_col = "id",
  code_col = "icd_code",
  log_path = NULL
)
}
\arguments{
\item{data}{Data frame containing pre-processed diagnostic data (check minimum requirements in documentation)}

\item{codes}{Character vector including ICD-10 codes to validate and filter diagnostic data}

\item{pattern_codes}{Character vector including general pattern of desired codes. For example: F84 will return all codes starting with F84.}

\item{id_col}{Name of ID column in data set, default is "id"}

\item{code_col}{String containing the name of column containing the diagnostic codes}

\item{log_path}{File path of the log file to be used}
}
\value{
Filtered diagnostic data frame containing only relevant observations based on diagnostic codes of interest.
}
\description{
Validate and filter diagnostic data by selected ICD-10 codes
}
\examples{
# Validate that F45 and F84 are real codes/family of codes in ICD-10.
# Keep only rows with codes containing F45 and F84.

log_file <- tempfile()
cat("Example log file", file = log_file)

filtered_diag_df <-  filter_diag(data = diag_df,
                                 pattern_codes = c("F45", "F84"),
                                 id_col = "id",
                                 code_col = "code",
                                 log_path = log_file
                                 )

}
