% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_diag_codes.R
\name{filter_diag}
\alias{filter_diag}
\title{Validate and filter diagnostic data}
\usage{
filter_diag(data, codes, code_col = "icd_code")
}
\arguments{
\item{data}{Data frame containing pre-processed diagnostic data (check minimum requirements in documentation)}

\item{codes}{Character vector including ICD-10 codes to validate and filter diagnostic data}

\item{code_col}{String containing the name of column containing the diagnostic codes}
}
\value{
Filtered diagnostic data frame containing only relevant observations based on diagnostic codes of interest.
}
\description{
Validate and filter diagnostic data
}
