% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.effect.sizes}
\alias{get.effect.sizes}
\title{Simple function to compute effect sizes used by effectplot}
\usage{
get.effect.sizes(reg, dat, vars = intersect(colnames(dat), names(coef(reg))),
  scale.depvar = NULL, depvar = names(reg$model)[[1]], data.fun = NULL,
  numeric.effect = "10-90", dummy01 = TRUE, predict.type = "response",
  effect.bases = NULL, compute.se = FALSE, ci.prob = 0.95,
  repl.for.se = 10000)
}
\description{
Simple function to compute effect sizes used by effectplot
}

