% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtools.R
\name{get.regression.data}
\alias{get.regression.data}
\title{Extract data from a regression model}
\usage{
get.regression.data(mod, source.data = NULL, ...)
}
\description{
First looks whether the model mod has a field 'custom.data' in which a user can manually store the data.frame.
}

