% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.regression.var}
\alias{get.regression.var}
\title{Extract values of the variable var from a regression model mod}
\usage{
get.regression.var(mod, var, as.data.frame = length(var) > 1)
}
\description{
Reports an error if the data for that variable cannot get extracted
from the regression model mod.
}

