% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_ssb.R
\name{get_population_ssb}
\alias{get_population_ssb}
\title{Get yearly population data from SSB, by region, sex and age.}
\usage{
get_population_ssb(
  url_api = "https://data.ssb.no/api/v0/en/table/07459/",
  regions = c("norway", "fylker", "kommuner"),
  years,
  ages,
  aggregate_age = TRUE,
  by_sex = TRUE,
  save_xslx = FALSE
)
}
\arguments{
\item{url_api}{Character string. URL used for the SSB API call. \emph{Do not modify}.}

\item{regions}{Character string. Supported regions:
\itemize{
\item "norway" for the country-wise population counts
\item "fylker" for county population counts
\item "kommuner" for municipality population counts. Due to region reforms in Norway, the results are shown using SSB's harmonized county codes (2024)
}}

\item{years}{Numerical vector. Year(s) of population counts.}

\item{ages}{Numerical vector. Age(s) in whole years.}

\item{aggregate_age}{Logical. Default is \code{TRUE}.
\itemize{
\item If \code{TRUE} and more than one age given, include new row with aggregated population for age group.
}}

\item{by_sex}{Logical. Default is \code{TRUE}.
\itemize{
\item If \code{TRUE} population counts are disaggregated by sex (male, female)
}}

\item{save_xslx}{Logical. Want to save the results as a xslx? Default is \code{FALSE}.
\itemize{
\item If \code{TRUE} results are saved as xslx in the current working directory.
}}
}
\value{
Data frame with population data from SSB
}
\description{
Get yearly population data from SSB, by region, sex and age.
}
\examples{
# Population of Norway in 2020 to 2022, for ages 10 to 15,
# include total of that age group (10-15) for each year.
population_norway <- get_population_ssb(regions = "norway",
                                        years = c(2020:2022),
                                        ages = c(10:15),
                                        aggregate_age = TRUE,
                                        by_sex = TRUE,
                                        save_xslx = FALSE)

}
