% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glm.marginal.effects}
\alias{glm.marginal.effects}
\title{Wrapper to mfx package to extract marginal effects from a glm model}
\usage{
glm.marginal.effects(mod, mfx.fun = NULL, ...)
}
\description{
not yet all possible types of glm models implemented, need to extend
}

