% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_municipality_codes.R
\name{harmonize_municipality_codes}
\alias{harmonize_municipality_codes}
\title{Harmonize old municipality codes in Norway to 2024 municipality codes}
\usage{
harmonize_municipality_codes(data, municipality_col = "code", fylke = TRUE)
}
\arguments{
\item{data}{Data containing old municipality codes}

\item{municipality_col}{Name of column containing original municipality codes}

\item{fylke}{Also want fylke name in 2024?}
}
\value{
Harmonized codes in 2024
}
\description{
Harmonize old municipality codes in Norway to 2024 municipality codes
}
\examples{
# Harmonize municipality codes from 2016 to 2024

harmonized_codes <- harmonize_municipality_codes(data = kommuner_2016, municipality_col = "code")
}
