% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_diag_demo.R
\name{link_diag_demo}
\alias{link_diag_demo}
\title{Link diagnostic and demographic datasets using unique personal identifiers}
\usage{
link_diag_demo(
  data_diag,
  data_demo_inv = NULL,
  data_demo_var = NULL,
  id_col = "id",
  date_col = "year",
  log_path = NULL
)
}
\arguments{
\item{data_diag}{Data frame containing pre-processed and pre-validated diagnostic data (check minimum requirements in documentations)}

\item{data_demo_inv}{Data frame containing validated time-invariant demographic data (check minimum requirements in documentations)}

\item{data_demo_var}{Data frame containing validated time-variant demographic data (check minimum requirements in documentations)}

\item{id_col}{Name (character) of the ID column in the data set (unique personal identifier), default is "id"}

\item{date_col}{Name (character) of the date column in the data set, in case of using time-variant data.}

\item{log_path}{File path of the log file to be used}
}
\value{
Linked dataset including relevant diagnostic and demographic characteristics.
}
\description{
Link diagnostic and demographic datasets using unique personal identifiers
}
\examples{
# Link diagnostic and time invariant datasets
log_file <- tempfile()
cat("Example log file", file = log_file)

linked_diag_inv <- link_diag_demo(data_diag = diag_df,
                                  data_demo_inv = invar_df,
                                  id_col = "id",
                                  log_path = log_file)

# Link diagnostic and time variant datasets
names(var_df)[names(var_df) == 'year_varying'] <- 'year'
names(diag_df)[names(diag_df) == 'diag_year'] <- 'year'

linked_diag_var <- link_diag_demo(data_diag = diag_df,
                                  data_demo_var = var_df,
                                  id_col = "id",
                                  date_col = "year",
                                  log_path = log_file)

# Link diagnostic, time invariant and variant datasets
linked_diag_inv_var <- link_diag_demo(data_diag = diag_df,
                                      data_demo_var = var_df,
                                      data_demo_inv = invar_df,
                                      id_col = "id",
                                      date_col = "year",
                                      log_path = log_file)

}
