% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_diag_demo.R
\name{link_diag_demo}
\alias{link_diag_demo}
\title{Link diagnostic and demographic datasets using unique personal identifiers}
\usage{
link_diag_demo(
  data_diag,
  data_demo_inv = NULL,
  data_demo_var = NULL,
  id_col = "id",
  date_col = "year"
)
}
\arguments{
\item{data_diag}{Data frame containing pre-processed and pre-validated diagnostic data (check minimum requirements in documentations)}

\item{data_demo_inv}{Data frame containing validated time-invariant demographic data (check minimum requirements in documentations)}

\item{data_demo_var}{Data frame containing validated time-variant demographic data (check minimum requirements in documentations)}

\item{id_col}{Name (character) of the ID column in the data set (unique personal identifier), default is "id"}

\item{date_col}{Name (character) of the date column in the data set, in case of using time-variant data.}
}
\value{
Linked dataset including relevant diagnostic and demographic characteristics.
}
\description{
Link diagnostic and demographic datasets using unique personal identifiers
}
