\name{misc}
\alias{replicMeans}

\title{Utilities}

\description{
Various helper functions.
}

\usage{
replicMeans(nrep,toReplic,timing=FALSE)
}

\arguments{
  \item{nrep}{Number of replications.}
  \item{toReplic}{Function call(s), as a quoted string, separated by
     semicolons if more than one call.}
}

\details{
}

\examples{

}

\author{
Norm Matloff
}

