% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normalize.column.names}
\alias{normalize.column.names}
\title{Normalize column names such that they are valid variable names in R}
\usage{
normalize.column.names(names)
}
\description{
Normalize column names such that they are valid variable names in R
}

