\name{Penrose Linear}
\alias{penroseLM}
\alias{penrosePoly}
\alias{predict.penroseLM}
\alias{predict.penrosePoly}

\title{Penrose-Inverse Linear Models}

\description{
UNDER CONSTRUCTION

Provides mininum-norm solutions to linear models, identical to OLS in
standard situations, but allowing exploration of overfitting in the
overparameterized case. Also provides a convenient wrapper for polynomial
regression problems.
}

\usage{
}

\arguments{
}

\details{

}

\value{

}

\examples{

}

\author{
Norm Matloff
}

