% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rates.R
\name{plot_rates}
\alias{plot_rates}
\title{Title}
\usage{
plot_rates(
  data,
  date_col,
  rate_col,
  grouping_var = NULL,
  facet_var = NULL,
  plot_type = c("line", "bar_chart", "lollipop", "jitter"),
  percent = TRUE,
  palette = c("fhi_colors", "viridis", "okabe_ito"),
  single_color = "black",
  annotate_line = NULL,
  CI_lower = NULL,
  CI_upper = NULL,
  plot_title = "",
  x_name = "",
  y_name = "",
  legend_title = "",
  coord_flip = F,
  start_end_points = F,
  interactive = FALSE
)
}
\arguments{
\item{coord_flip}{}
}
\description{
Title
}
