% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_series.R
\name{prevalence_series}
\alias{prevalence_series}
\title{Calculate prevalence series rates}
\usage{
prevalence_series(
  linked_data,
  time_points,
  id_col = "id",
  date_col = "date",
  pop_data,
  pop_col = "pop_count",
  grouping_vars = NULL,
  only_counts = FALSE,
  suppression = TRUE,
  suppression_treshold = 5
)
}
\arguments{
\item{linked_data}{Dataset containing relevant diagnostic and demographic information}

\item{time_points}{Time points.}

\item{id_col}{Name (character) of the ID column in the data set (unique personal identifier). Default is "id".}

\item{date_col}{Name (character) of the date column in the data set. Default is "date".}

\item{pop_data}{Dataset containing relevant population information.}

\item{pop_col}{Name (character) of the column containing population counts in the population dataset.}

\item{grouping_vars}{Optional character vector including grouping variables for the aggregation of diagnostic counts (eg. sex, education).}

\item{only_counts}{Return only diagnostic count, instead of prevalence rates. Default is set to FALSE.}

\item{suppression}{Apply suppression to results (intermediate and rates) in order to maintain statistical confidentiality.}

\item{suppression_treshold}{Threshold for suppression, default is set to 5 (NPR standard).}
}
\value{
Prevalence rate rates for specified time points
}
\description{
Calculate prevalence series rates
}
