% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_demo_data.R
\name{read_demo_data}
\alias{read_demo_data}
\title{Read and validate the structure of demographic individual-level data}
\usage{
read_demo_data(
  file_path,
  data_type = c("t_variant", "t_invariant"),
  id_col = "id",
  date_col = "date",
  log_path = NULL,
  ...
)
}
\arguments{
\item{file_path}{A character string. File path to the demographic data to read. Supports CSV, RDS, RDA and .SAV files.}

\item{data_type}{A character string. Demographic data can either be of type "t_variant" or "t_invariant", necessary to check correct data structure characteristics.}

\item{id_col}{A character string. Name of ID column in data set. Default is "id".}

\item{date_col}{A character string. Name of date column in data set, default is "date".}

\item{log_path}{A character string. Path to the log file to append function logs. Default is \code{NULL}.
\itemize{
\item If \code{NULL}, a new directory \verb{/log} and file is created in the current working directory.
}}

\item{...}{Additional arguments passed to methods or underlying functions.}
}
\value{
A data frame with the validated minimum requirements for demographic data.
}
\description{
\code{read_demo_data()} validates the general structure and minimum column requirements for demographic individual-level data.
The input data sets must be CSV, RDS, RDA or .SAV files.
}
\examples{
# Read and validate CSV file for varying individual level demographic data
demo_csv <- system.file("extdata", "invar_data.csv", package = "regtools")
log_file <- tempfile()
cat("Example log file", file = log_file)

demo_data_validated <- read_demo_data(demo_csv, data_type = "t_invariant",
id_col = "id", log_path = log_file)

}
