% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_diagnostic.R
\name{read_diag_data}
\alias{read_diag_data}
\title{Read and validate the structure of diagnostic data}
\usage{
read_diag_data(
  file_path,
  id_col = "id",
  date_col = "date",
  code_col = "code",
  ...
)
}
\arguments{
\item{file_path}{File path of the diagnostic data to be read. Supports CSV, RDS, RDA and .SAV files.}

\item{id_col}{Name of ID column in data set, default is "id"}

\item{date_col}{Name of date column in data set, default is "date"}

\item{code_col}{Name of diagnostic codes column in data set, default is "code"}

\item{...}{Optional extra parameters for specifying correct reading of CSV and .SAV files}
}
\value{
A data frame with the validated minimum requirements for diagnostic data
}
\description{
`read_data_diagnostic()` validates the general structure and minimum column requirements for diagnostic data.
The input data sets must be CSV, RDS, RDA or .SAV files.
}
