% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtools.R
\name{regression.data}
\alias{regression.data}
\title{Creates a data.frame that can be used to run a regression with the given formula}
\usage{
regression.data(formula = stats::formula(reg), reg = NULL, data = NULL,
  normalize.names = TRUE, remove.intercept = TRUE, expand = FALSE)
}
\arguments{
\item{formula}{a regression formula e.g. y~log(x1)+x2}

\item{reg}{optionally a regression object from which the formula is takes}

\item{data}{a data frame that contains original variables used in formula}

\item{normalize.names}{if TRUE (default) change column names to be valid R variable names}

\item{remove.intercept}{if TRUE (default) do not generate an .Intercept column}

\item{expand}{if TRUE, call expanded.regression.data instead}
}
\value{
a data.frame
}
\description{
Creates a data.frame that can be used to run a regression with the given formula
}

