% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtools-package.R
\docType{package}
\name{regtools-package}
\alias{regtools}
\alias{regtools-package}
\title{regtools: An R package to aid the analysis of Norwegian individual-level health and sociodemographic data.}
\description{
R package to support research activities utilizing Norwegian health records for prevalence and trend analysis of mental health diagnoses in Norway.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/amslala/regtools}
  \item \url{https://amslala.github.io/regtools/}
  \item Report bugs at \url{https://github.com/amslala/regtools/issues}
}

}
\author{
\strong{Maintainer}: Alejandra Martinez Sanchez

Other contributors:
\itemize{
  \item Johanne Pettersen
  \item Helga Ask
  \item Alexandra Havdahl
  \item Laurie Hannigan
}

}
\keyword{internal}
