% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtools-package.R
\docType{package}
\name{regtools-package}
\alias{regtools}
\alias{regtools-package}
\title{regtools: regtools: Trend/prevalence analysis and visualization tool for mental health diagnoses in Norwegian health registers}
\description{
R package to support research activities utilizing Norwegian health records for prevalence and trend analysis of mental health diagnoses in Norway.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/amslala/regtools}
  \item Report bugs at \url{https://github.com/amslala/regtools/issues}
}

}
\keyword{internal}
