
\name{regtools-package}
\alias{regtools-package}
\alias{regtools}
\docType{package}         

\title{Overview and Package Reference Guide}
                                                                                \description{ 

This package provides a broad collection of functions useful for
regression and classification analysis.

}

\details{

   \itemize{

   \item k-NN: kmin, knnest, knntrn, preprocessx, meany, vary, loclin,
   predict, kmin, pwplot

   \item classification:  classadjust, avalogtrn, avalogpred, ovalogtrn,
   ovalogpred, knntrn, predict, pwplot, nonparvsxplot

   \item regression diagnostics:  parvsnonparplot, nonparvsxplot,
   nonparvarplot

   \item nonlinear regression:  nlshc

   \item ridge regression:  ridgelm, plot

   \item conversion between factors and dummies: dummiesToFactor,
   factorsToDummies, factorToDummies

   \item misc.:  mm, tabletofakedf, unscale
   }

}

