% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustSE.r
\name{robust.se.summary}
\alias{robust.se.summary}
\title{Show a model summary with robust standard errors
 
Using the sandwich package, creates summary with heteroscedasticity consistent standard errors (Stata´s robust dafault (Stata (2014)))}
\usage{
robust.se.summary(m, type = "HC3", ...)
}
\description{
Show a model summary with robust standard errors
 
Using the sandwich package, creates summary with heteroscedasticity consistent standard errors (Stata´s robust dafault (Stata (2014)))
}

