% Generated by roxygen2 (4.0.1): do not edit by hand
\name{showreg}
\alias{showreg}
\title{An extension of screenreg, texreg and htmlreg that allows for robust standard errors.}
\usage{
showreg(l, custom.model.names = names(l), robust = FALSE,
  robust.type = "HC3", cluster1 = NULL, cluster2 = NULL,
  coef.transform = NULL, vcov.li = NULL, coef.mat.li = NULL, digits = 2,
  output = c("text", "html", "latex"), output.fun = NULL, doctype = FALSE,
  ...)
}
\arguments{
\item{custom.model.names}{custom titles for each model}

\item{robust}{shall robust standard errors be used? Logical single number or a vector specifying for each model.}

\item{robust.type}{the type of robust standard errors. Can be "HAC", "cluster" or "HC1"-"HC4"}

\item{cluster1}{and cluster2 if clustered robust standard errors are computed the name of the variables that shall be clustered by}

\item{coef.transform}{either NULL or a vector containing "no" or "mfx", if an entry is "mfx" we show the marginal effects of the corresponding model.}

\item{...}{additional parameters for screenreg, texreg or htmlreg}
}
\description{
An extension of screenreg, texreg and htmlreg that allows for robust standard errors.
}

