% Generated by roxygen2 (4.0.1): do not edit by hand
\name{showreg}
\alias{showreg}
\title{An extension of screenreg, texreg and htmlreg that allows for robust standard errors.}
\usage{
showreg(l, custom.model.names = names(l), robust = FALSE,
  robust.type = "HC3", cluster1 = NULL, cluster2 = NULL,
  coef.transform = NULL, vcov.li = NULL, coef.mat.li = NULL, digits = 2,
  output = c("text", "html", "latex"), output.fun = NULL, doctype = FALSE,
  ...)
}
\arguments{
\item{custom.model.names}{custom titles for each model}

\item{robust}{shall robust standard errors be used? Logical single number or a vector specifying for each model.}

\item{robust.type}{the type of robust standard errors. Can be "HAC", "cluster" or "HC1"-"HC4"}

\item{cluster1}{and cluster2 if clustered robust standard errors are computed the name of the variables that shall be clustered by}

\item{coef.transform}{either NULL or a vector containing "no" or "mfx", if an entry is "mfx" we show the marginal effects of the corresponding model.}

\item{...}{additional parameters for screenreg, texreg or htmlreg}
}
\description{
An extension of screenreg, texreg and htmlreg that allows for robust standard errors.
}

\examples{\dontrun{
  

  # iv and ols with robust standard errors
  library(AER)
  data("CigarettesSW", package = "AER")
  CigarettesSW$rprice <- with(CigarettesSW, price/cpi)
  CigarettesSW$rincome <- with(CigarettesSW, income/population/cpi)
  CigarettesSW$tdiff <- with(CigarettesSW, (taxs - tax)/cpi)

  iv <- ivreg(log(packs) ~ log(rprice) + log(rincome) | log(rincome) + tdiff + I(tax/cpi),data = CigarettesSW, subset = year == "1995")
  ols <- lm(log(packs) ~ log(rprice) + log(rincome),data = CigarettesSW, subset = year == "1995")

  showreg(list(iv=iv,iv.rob=iv, ols=ols,  ols.rob=ols),
          robust=c(FALSE,TRUE,FALSE,TRUE), robust.type="HC4")  


  # Marginal effect for probit regression
  set.seed(12345)
  n = 1000
  x = rnorm(n)
  
  # binary outcome
  y = ifelse(pnorm(1 + 4*x + rnorm(n))>0.5, 1, 0)
  data = data.frame(y,x)

  reg = glm(y~x, data=data, family=binomial(link=probit))
  showreg(list("probit"=reg,"marginal effects"=reg), coef.transform=c("no", "mfx"), omit.coef = "(Intercept)")

  # Clustered standard errors
  # (not tested for correctness at all)
  library(Ecdat)
  data(Fatality)
  LSDV <- lm(mrall ~ beertax + factor(year) + factor(state), data=Fatality)
  LSDV$custom.data = Fatality
  showreg(
    list("LSDV"=LSDV,
         "LSDV (state cluster)"=LSDV,
         "LSDV (state-year cluster)"=LSDV
    ),
    robust=c(FALSE,TRUE,TRUE),
    robust.type=c("const","cluster1","cluster2"),
    cluster1 = "state", cluster2="year"
  )

}}
