% Generated by roxygen2 (4.0.1): do not edit by hand
\name{showreg}
\alias{showreg}
\title{Show and compare regression results}
\usage{
showreg(l, custom.model.names = names(l), robust = FALSE,
  robust.type = "HC3", cluster1 = NULL, cluster2 = NULL, vcov.li = NULL,
  coef.transform = NULL, coef.mat.li = NULL, digits = 2,
  output = c("text", "html", "latex"), output.fun = NULL, doctype = FALSE,
  ...)
}
\arguments{
\item{l}{list of models as in screenreg}

\item{custom.model.names}{custom titles for each model. By default the names of the model list.}

\item{robust}{shall robust standard errors be used? Logical single number or a vector specifying for each model.}

\item{robust.type}{the type of robust standard errors. Can be "HAC", "cluster", "HC1" to "HC4" or "NeweyWest". Can be a vector specifying a type for each model.}

\item{cluster1}{and cluster2 if clustered robust standard errors are used, the name of the variables that shall be clustered by}

\item{vcov.list}{optional a list of covariance matrices of the coefficients for every model}

\item{coef.transform}{either NULL or a vector containing "no" or "mfx", if an entry is "mfx" we show the marginal effects of the corresponding model.}

\item{coef.mat.li}{for highest flexibility, you can also provide a list that contains for each model a matrix or data.frame as returned by coeftest with the columns: coefficent, se, t-value, p-value.}

\item{output}{either "text", "html" or "latex"}

\item{output.fun}{allows a manual output function, e.g. if one has overloaded the design of screenreg, texreg or htmlreg.}

\item{...}{additional parameters for screenreg, texreg or htmlreg}
}
\description{
The function extends and wraps the screenreg, texreg and htmlreg functions in the texreg package. It allows for robust standard errors (also clustered robust standard errors) and can show marginal effects in glm models.
}

\examples{\dontrun{
  

  # iv and ols with robust standard errors
  library(AER)
  data("CigarettesSW", package = "AER")
  CigarettesSW$rprice <- with(CigarettesSW, price/cpi)
  CigarettesSW$rincome <- with(CigarettesSW, income/population/cpi)
  CigarettesSW$tdiff <- with(CigarettesSW, (taxs - tax)/cpi)

  iv <- ivreg(log(packs) ~ log(rprice) + log(rincome) | log(rincome) + tdiff + I(tax/cpi),data = CigarettesSW, subset = year == "1995")
  ols <- lm(log(packs) ~ log(rprice) + log(rincome),data = CigarettesSW, subset = year == "1995")

  showreg(list(iv=iv,iv.rob=iv, ols=ols,  ols.rob=ols),
          robust=c(FALSE,TRUE,FALSE,TRUE), robust.type="HC4")  
  showreg(list(iv=iv,iv.rob=iv, ols=ols,  ols.rob=ols),
          robust=c(FALSE,TRUE,FALSE,TRUE), robust.type="NeweyWest", prewhite=FALSE)  


  # Marginal effect for probit regression
  set.seed(12345)
  n = 1000
  x = rnorm(n)
  
  # binary outcome
  y = ifelse(pnorm(1 + 4*x + rnorm(n))>0.5, 1, 0)
  data = data.frame(y,x)

  reg = glm(y~x, data=data, family=binomial(link=probit))
  showreg(list("probit"=reg,"marginal effects"=reg), coef.transform=c("no", "mfx"), omit.coef = "(Intercept)")

  # Clustered standard errors
  # (not tested for correctness at all)
  library(Ecdat)
  data(Fatality)
  LSDV <- lm(mrall ~ beertax + factor(year) + factor(state), data=Fatality)
  LSDV$custom.data = Fatality
  showreg(
    list("LSDV"=LSDV,
         "LSDV (state cluster)"=LSDV,
         "LSDV (state-year cluster)"=LSDV
    ),
    robust=c(FALSE,TRUE,TRUE),
    robust.type=c("const","cluster1","cluster2"),
    cluster1 = "state", cluster2="year"
  )

}}
