\name{unscale}
\alias{unscale}

\title{Miscellaneous Utilities}

\description{
Utilities.
}

\usage{
unscale(scaledx)
}

\arguments{
  \item{scaledx}{Object returned from \code{scale}.}
}

\value{
The function \code{unscale} returns the original object to which
\code{scale} had been applied.
}

\author{
Norm Matloff
}

