\name{unscale}
\alias{unscale}

\title{Miscellaneous Utilities}

\description{
Utilities.
}

\usage{
unscale(scaledx,ctrs=NULL,sds=NULL)
}

\arguments{
  \item{scaledx}{Matrix.}
}

\value{
The function \code{unscale} returns the original object to which
\code{scale} had been applied.  Or, the attributes \code{ctrs} and
\code{sds} can be specified by the user.
}

\author{
Norm Matloff
}

