% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustSE.r
\name{vcovCluster}
\alias{vcovCluster}
\title{Get clustered robust standard errors for regression model mod}
\usage{
vcovCluster(mod, cluster1, cluster2 = NULL, dfcw = 1)
}
\arguments{
\item{cluster1}{variable name of cluster1}

\item{cluster2}{variable name of cluster2}
}
\description{
R-codes (www.r-project.org) for computing clustered-standard errors from Mahmood Arai, Jan 26, 2008.
}

